--- src/main/java/Soon/Coming/springbootdeveloper/post/controller/PostController.java
@@ -1,25 +1,0 @@
-package Soon.Coming.springbootdeveloper.post.controller;
-
-import Soon.Coming.springbootdeveloper.post.domain.Post;
-import Soon.Coming.springbootdeveloper.post.dto.CreatePostRequest;
-import Soon.Coming.springbootdeveloper.post.service.PostService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Service;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RestController
-public class PostController {
-
-    private final PostService postService;
-
-    //api post
-    @PostMapping("/posts")
-    public ResponseEntity<Post> createPost(@RequestBody @Validated CreatePostRequest createPostRequest){
-        return ResponseEntity.ok().body(postService.createPost(createPostRequest));
-    }
-}
