+++ src/main/java/Soon/Coming/springbootdeveloper/post/dto/ArticleResponse.java
@@ -1,14 +1,14 @@
-package Soon.Coming.springbootdeveloper.post_crud.dto;
+package Soon.Coming.springbootdeveloper.post.dto;
 
-import Soon.Coming.springbootdeveloper.post_crud.domain.Article;
 import lombok.Getter;
+import Soon.Coming.springbootdeveloper.post.domain.Article;
+
 
 @Getter
 public class ArticleResponse {
 
     private final String title;
     private final String content;
-
     public ArticleResponse(Article article) {
         this.title = article.getTitle();
         this.content = article.getContent();
