+++ src/main/java/Soon/Coming/springbootdeveloper/post/repository/PostRepository.java
@@ -1,11 +1,7 @@
 package Soon.Coming.springbootdeveloper.post.repository;
 
-import Soon.Coming.springbootdeveloper.post.domain.Post;
+import Soon.Coming.springbootdeveloper.post.domain.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-
-@Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
-
+public interface PostRepository extends JpaRepository<Article, Long> {
 }
