+++ src/main/java/user/config/WebSecurityConfig.java
@@ -1,6 +1,6 @@
-package Soon.Coming.springbootdeveloper.config;
+package user.config;
 
-import Soon.Coming.springbootdeveloper.service.UserDetailService;
+import user.service.UserDetailService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -23,7 +23,6 @@ import static org.springframework.boot.autoconfigure.security.servlet.PathReques
 public class WebSecurityConfig {
 
     private final UserDetailService userService;
-
 
     @Bean
     /*1. 스프링 시큐리티 기능 비활성화*/
