+++ src/main/java/user/domain/User.java
@@ -1,6 +1,5 @@
-package Soon.Coming.springbootdeveloper.domain;
+package user.domain;
 
-import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,7 +12,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import java.util.Collection;
 import java.util.List;
 
-@Table(name = "users") // 작은따옴표('') 대신 큰따옴표("") 사용
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
