+++ src/main/java/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package Soon.Coming.springbootdeveloper.repository;
+package user.repository;
 
-import Soon.Coming.springbootdeveloper.domain.User;
+import user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
