+++ src/main/java/user/service/UserDetailService.java
@@ -1,14 +1,14 @@
-package Soon.Coming.springbootdeveloper.service;
+package user.service;
 
-import Soon.Coming.springbootdeveloper.domain.User;
-import Soon.Coming.springbootdeveloper.repository.UserRepository;
+
+import user.domain.User;
+import user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
 @Service
-//JSDoc 처음 써보는데 이거이거 아주 깔@롱한데? 멋져
 
 //스프링 시큐리티에서 사용자 정보를 가져오는 인터페이스
 public class UserDetailService implements UserDetailsService {
